Add files via upload
This is the first outcast of a style sheet with corresponding source files. The xml and yml files have code that refers to files that are assumed to be in the same folder - hence al these files need to be kept together and the source folder (DARE.tm2source) needs to be within the style folder (DARE_style_2.tm2).